pyproject.toml
[tool.poetry]
name = "movies"
version = "0.1.0"
description = "Data pipeline for NYT movie reviews"
authors = ["Alexander Butz <alex.butz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
apache-airflow = "^2.3.4"

[tool.poetry.dev-dependencies]
bpython = "^0.23"
pre-commit = "^2.20.0"
darglint = "^1.8.1"
flake8 = "^5.0.4"
flake8-docstrings = "^1.6.0"
isort = "^5.10.1"
mypy = "^0.971"
pep8-naming = "^0.13.2"
pydocstyle = "^6.1.1"
black = "^22.8.0"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
